Automated Endmember Selection is going poorly
Having pretty persistent issues correctly selecting endmembers for the sentinel data. The only difference between the input data for sentinel and landsat is the resolution, and that landsat has a panchromatic band, where sentinel doesn't. The landsat outputs look pretty good to me, but the sentinel ones are garbage, producing endmembers that yield a homogeneous classification across the image.